Move builder and knowledge base to separate /kb folder, and adapt examples
