+++ src/legacy/tests/legacy_test_tke_client.py
@@ -1,4 +1,4 @@
-import src.knowledge_base as tke_kb
+import src.kb.knowledge_base as tke_kb
 import src.knowledge_interaction as tke_ki
 import src.tke_client as tke
 import pytest
